+++ src/config/chain/stargaze.ts
@@ -45,10 +45,6 @@ export const STARGAZE_CHAIN_CONFIGS: ChainConfigs = {
       gasAdjustment: 1.5,
       maxGasLimit: 25_000_000,
     },
-    explorerLink: {
-      validator: "https://www.mintscan.io/stargaze/validators",
-      proposal: "https://www.mintscan.io/stargaze/proposals",
-    },
     extra: {
       singleStakingDenom: "ustars",
     },
@@ -94,10 +90,6 @@ export const STARGAZE_CHAIN_CONFIGS: ChainConfigs = {
       },
       gasAdjustment: 1.5,
       maxGasLimit: 50_000_000,
-    },
-    explorerLink: {
-      validator: "https://testnet-explorer.publicawesome.dev/stargaze/staking",
-      proposal: "https://testnet-explorer.publicawesome.dev/stargaze/gov",
     },
     extra: {
       singleStakingDenom: "ustars",